+++ R/extend_to_exergy.R
@@ -287,6 +287,13 @@ move_to_exergy <- function(psut_energy,
   # Make sure we're operating on the countries of interest.
   psut_energy <- psut_energy %>%
     dplyr::filter(.data[[country]] %in% countries)
+  # If the psut_energy data frame has no rows, 
+  # simply return it. 
+  # Both the incoming and outgoing data frames have the exact same columns.
+  if (nrow(psut_energy) == 0) {
+    return(psut_energy)
+  }
+  # We have a non-zero number of rows, so proceed with the calculations.
   phi_vecs <- phi_vecs %>%
     dplyr::filter(.data[[country]] %in% countries)