+++ _targets.R
@@ -19,6 +19,11 @@ years <- 1960:2019                         # The years to be analyzed
 #                "GRC", "HKG", "HND", "IDN", "IND", "JOR", "JPN", "KOR", "MEX",
 #                "NOR", "PRT", "RUS", "USA", "WABK", "WMBK", "ZAF")
 
+# countries <- c("BRA", "CAN", "CHNM", "DEU", "DNK", "ESP", "FRA", "GBR", "GHA",
+#                "GRC", "HKG", "HND", "IDN", "IND", "JOR", "JPN", "KOR", "MEX",
+#                "NOR", "PRT", "RUS", "USA", "ZAF")
+
+# countries <- c("WABK", "GBR")
 
 countries <- c("GBR")